Deduplicate common methods for the updaters
