+++ app/jobs/updater.rb
@@ -1,4 +1,16 @@
 class Updater < ApplicationJob
+  attr_reader :importer
+
+  def perform()
+    import
+    pre_grouper = PreGrouper.new
+    pre_grouper.perform
+    grouper = Grouper.new
+    grouper.perform(should_group_genes?, should_group_drugs?, importer.source.id)
+    post_grouper = PostGrouper.new
+    post_grouper.perform(should_cleanup_gene_claims?)
+  end
+
   def should_group_genes?
     raise StandardError.new('Must implement #should_group_genes? in subclass')
   end
@@ -9,5 +21,9 @@ class Updater < ApplicationJob
 
   def should_cleanup_gene_claims?
     false
+  end
+
+  def create_importer
+    raise StandardError.new('Must implement #create_importer in subclass')
   end
 end
